fix: [#312] Review 반영
